Total order in AABBTreePointsMaker::partitionPoints
