+++ source/MRTest/MRTest.vcxproj.filters
@@ -103,6 +103,9 @@
     <ClCompile Include="MRFeaturesTests.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="MRAABBTreePointsTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\.editorconfig" />
